fix: fix queries
